refactor: move some types around
