+++ src/app/api/apiHandler.ts
@@ -1,7 +1,8 @@
 import 'server-only'
 import { Session } from '@/auth/Session'
 import { ServerError, Smorekopp } from '@/services/error'
-import type { ServiceMethodReturn, Validation } from '@/services/ServiceMethod'
+import type { ValidationTypeUnknown } from '@/services/Validation'
+import type { ServiceMethodType } from '@/services/ServiceMethod'
 import type { ErrorCode } from '@/services/error'
 import type { SessionNoUser } from '@/auth/Session'
 import type { z } from 'zod'
@@ -10,9 +11,9 @@ type APIHandler<
     RawParams,
     Return,
     ParamsSchema extends z.ZodTypeAny | undefined,
-    DataValidation extends Validation<unknown, unknown> | undefined,
+    DataValidation extends ValidationTypeUnknown | undefined,
 > = {
-    serviceMethod: ServiceMethodReturn<boolean, Return, ParamsSchema, DataValidation>,
+    serviceMethod: ServiceMethodType<boolean, Return, ParamsSchema, DataValidation>,
 } & (ParamsSchema extends undefined ? {
     params?: undefined,
 } : {
@@ -40,7 +41,7 @@ export function apiHandler<
     RawParams,
     Return,
     ParamsSchema extends z.ZodTypeAny | undefined = undefined,
-    DataValidation extends Validation<unknown, unknown> | undefined = undefined,
+    DataValidation extends ValidationTypeUnknown| undefined = undefined,
 >({ serviceMethod, params }: APIHandler<RawParams, Return, ParamsSchema, DataValidation>) {
     return async (req: Request, { params: rawParams }: { params: RawParams }) =>
         await apiHandlerGeneric<Return>(req, async session => {
